Uploaded Version 2.1 of Windows 10 Cracker
this version was improved and is my first version that is being uploaded to github. the older versions are private or distributed by me on youtube.
This version changes a little bit of string to make this work with te latest versions! for the people wanting the older versions, they are coming soon!